requirements.txt
argon2-cffi>=23.1.0
argon2-cffi-bindings>=21.2.0
cffi>=1.15.1
pip>=22.0.4
pycparser>=2.21
setuptools>=58.1.0